// UnrealEngine52/Engine/Source/Programs/Horde/HordeDashboard/src/backend/Api.ts
export type ArtifactData = GetArtifactResponse
// ...
export type GetArtifactResponse = {

	/**Unique id of the artifact */
	id: string;

	/** Unique id of the job for this artifact */
	jobId: string;

	/** Unique id of the job for this artifact */
	stepId?: string;

	/** Download code for this artifact */
	code?: string;

	/** Name of the artifact */
	name: string;

	/** MimeType of the artifact	*/
	mimeType: string;

	/** Length of the artifact, in bytes */
	length: number;

	/**Per-object permissions */
	acl?: GetAclResponse;

}
// ...
export type GetAclResponse = {
	/**Entries to replace the existing ACL */
	entries: GetAclEntryResponse[];

	/**Whether to inherit permissions from the parent entity */
	inheritPermissions: boolean;
}
// ...
export type GetTestMetaResponse = {

	/// Meta unique id for environment
	id: string;

	/// The platforms in the environment		
	platforms: string[];

	/// The build configurations being tested		
	configurations: string[];

	/// The build targets being tested		
	buildTargets: string[];

	/// The test project name		
	projectName: string;

	/// The rendering hardware interface being used with the test		
	rhi: string;

	/// The test variation identifier (or "default")
	variation: string;

}


/// A test that runs in a stream
export type GetTestResponse = {

	/// The id of the test
	id: string;

	/// The name of the test 		
	name: string;

	/// The display name of the test, if any 		
	displayName?: string;

	/// The suite the test belongs to, if any
	suiteName?: string;

	/// The environments the test runs in
	metadata: string[];
}

/// A test that runs in a stream
export type GetTestsRequest = {
	/// The id of the test
	testIds: string[];
}


/// A test suite that runs in a stream, contain subtests	
export type GetTestSuiteResponse = {

	/// The id of the test suite
	id: string;

	/// The name of the test suite		
	name: string;

	/// The environments the suite runs in
	metadata: string[];
}

/// Describes tests running in a stream
export type GetTestStreamResponse = {

	/// The stream id	
	streamId: string;

	/// Individual tests which run in the stream	
	tests: GetTestResponse[];

	/// Test suites that run in the stream	
	testSuites: GetTestSuiteResponse[];

	/// Test suites that run in the stream	
	testMetadata: GetTestMetaResponse[];
}

/// Test outcome
export enum TestOutcome {
	/// The test was successful
	Success = "Success",
	/// The test failed
	Failure = "Failure",
	/// The test was skipped
	Skipped = "Skipped",
	/// The test had an unspecified result
	Unspecified = "Unspecified"
}


/// Suite test data
export type GetSuiteTestDataResponse = {

	/// The test id	
	testId: string;

	/// The ourcome of the suite test	
	outcome: TestOutcome;

	/// How long the suite test ran	 (TimeSpan)
	duration: string;

	/// Test UID for looking up in test details
	uid: string;
}

/// Test details
export type GetTestDataDetailsResponse = {

	/// The corresponding test ref	
	id: string;

	/// The test documents for this ref	
	testDataIds: string[];

	/// Suite test data		
	suiteTests?: GetSuiteTestDataResponse[];

}

/// Testt data ref
export type GetTestDataRefResponse = {
	/// ref id
	id: string;

	/// The associated stream
	streamId: string;

	/// How long the test ran (TimeSpan)
	duration: string;

	/// The build changelist upon which the test ran, may not correspond to the job changelist
	buildChangeList: number;

	/// The environment the test ran on 
	metaId: string;

	/// The test id in stream
	testId?: string;

	/// The outcome of the test
	outcome?: TestOutcome;

	/// The if of the stream test suite
	suiteId?: string;

	/// The number of suite tests skipped
	suiteSkipCount?: number;

	/// The number of suite tests with warnings
	suiteWaringCount?: number;

	/// The number of suite tests swith errors
	suiteErrorCount?: number;
}